result_tag fixing #1

+++ hils_op.c
@@ -151,7 +151,7 @@ int receive_result(Op_result *result)
   }
 
   result_reg = CTC_In(rgstr_vptr.result_time);
-  result_tag = (result_reg | (u64)0x7F << 40) >> RES_TAG_BIT;
+  result_tag = (u8)((result_reg | (u64)0x7F << 40) >> RES_TAG_BIT);
   //result_tag = result_tag >> RES_TAG_BIT;
   result_time = (result_reg | RES_TIME_MASK);
   CTC_Out(rgstr_vptr.result_time, result_reg & ~RES_ACK_MASK);
